--- a/src/OpenTelemetry.Exporter.Zipkin/Implementation/ZipkinSpan.cs
+++ b/src/OpenTelemetry.Exporter.Zipkin/Implementation/ZipkinSpan.cs
@@ -227,69 +227,69 @@ namespace OpenTelemetry.Exporter.Zipkin.Implementation
         {
             writer.WriteStartObject();
 
-            writer.WriteString("traceId", this.TraceId);
+            writer.WriteString(ZipkinSpanJsonHelper.TraceIdPropertyName, this.TraceId);
 
             if (this.Name != null)
             {
-                writer.WriteString("name", this.Name);
+                writer.WriteString(ZipkinSpanJsonHelper.NamePropertyName, this.Name);
             }
 
             if (this.ParentId != null)
             {
-                writer.WriteString("parentId", this.ParentId);
+                writer.WriteString(ZipkinSpanJsonHelper.ParentIdPropertyName, this.ParentId);
             }
 
-            writer.WriteString("id", this.Id);
+            writer.WriteString(ZipkinSpanJsonHelper.IdPropertyName, this.Id);
 
             if (this.Kind != null)
             {
-                writer.WriteString("kind", this.Kind);
+                writer.WriteString(ZipkinSpanJsonHelper.KindPropertyName, this.Kind);
             }
 
             if (this.Timestamp.HasValue)
             {
-                writer.WriteNumber("timestamp", this.Timestamp.Value);
+                writer.WriteNumber(ZipkinSpanJsonHelper.TimestampPropertyName, this.Timestamp.Value);
             }
 
             if (this.Duration.HasValue)
             {
-                writer.WriteNumber("duration", this.Duration.Value);
+                writer.WriteNumber(ZipkinSpanJsonHelper.DurationPropertyName, this.Duration.Value);
             }
 
             if (this.Debug.HasValue)
             {
-                writer.WriteBoolean("debug", this.Debug.Value);
+                writer.WriteBoolean(ZipkinSpanJsonHelper.DebugPropertyName, this.Debug.Value);
             }
 
             if (this.Shared.HasValue)
             {
-                writer.WriteBoolean("shared", this.Shared.Value);
+                writer.WriteBoolean(ZipkinSpanJsonHelper.SharedPropertyName, this.Shared.Value);
             }
 
             if (this.LocalEndpoint != null)
             {
-                writer.WritePropertyName("localEndpoint");
+                writer.WritePropertyName(ZipkinSpanJsonHelper.LocalEndpointPropertyName);
                 this.LocalEndpoint.Write(writer);
             }
 
             if (this.RemoteEndpoint != null)
             {
-                writer.WritePropertyName("remoteEndpoint");
+                writer.WritePropertyName(ZipkinSpanJsonHelper.RemoteEndpointPropertyName);
                 this.RemoteEndpoint.Write(writer);
             }
 
             if (!this.Annotations.IsEmpty)
             {
-                writer.WritePropertyName("annotations");
+                writer.WritePropertyName(ZipkinSpanJsonHelper.AnnotationsPropertyName);
                 writer.WriteStartArray();
 
                 foreach (var annotation in this.Annotations)
                 {
                     writer.WriteStartObject();
 
-                    writer.WriteNumber("timestamp", annotation.Timestamp);
+                    writer.WriteNumber(ZipkinSpanJsonHelper.TimestampPropertyName, annotation.Timestamp);
 
-                    writer.WriteString("value", annotation.Value);
+                    writer.WriteString(ZipkinSpanJsonHelper.ValuePropertyName, annotation.Value);
 
                     writer.WriteEndObject();
                 }
@@ -299,7 +299,7 @@ namespace OpenTelemetry.Exporter.Zipkin.Implementation
 
             if (!this.Tags.IsEmpty || this.LocalEndpoint.Tags != null)
             {
-                writer.WritePropertyName("tags");
+                writer.WritePropertyName(ZipkinSpanJsonHelper.TagsPropertyName);
                 writer.WriteStartObject();
 
                 // this will be used when we convert int, double, int[], double[] to string