Reuses the encoded property names to save some processing time when converting spans to Json in Zipkin.

## Before
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |      Error |     StdDev | Completed Work Items | Lock Contentions |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|------------------------ |---------------- |-------------- |------------:|-----------:|-----------:|---------------------:|-----------------:|-----------:|----------:|------:|----------:|
| ZipkinExporter_Batching |               1 |         10000 |    64.51 ms |   1.083 ms |   1.158 ms |               4.2500 |                - |   125.0000 |         - |     - |   1.71 MB |
| ZipkinExporter_Batching |              10 |         10000 |   600.62 ms |  10.988 ms |  10.278 ms |              24.0000 |                - |  2000.0000 |         - |     - |  16.95 MB |
| ZipkinExporter_Batching |             100 |         10000 | 6,360.48 ms | 124.514 ms | 224.524 ms |             205.0000 |                - | 21000.0000 | 2000.0000 |     - |  169.3 MB |

## After
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev | Completed Work Items | Lock Contentions |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|---------------------:|-----------------:|-----------:|----------:|------:|----------:|
| ZipkinExporter_Batching |               1 |         10000 |    56.29 ms |  0.752 ms |  0.703 ms |               4.2000 |                - |   200.0000 |         - |     - |   1.71 MB |
| ZipkinExporter_Batching |              10 |         10000 |   540.61 ms |  9.663 ms |  9.039 ms |              24.0000 |                - |  2000.0000 |         - |     - |  16.95 MB |
| ZipkinExporter_Batching |             100 |         10000 | 5,393.68 ms | 76.291 ms | 63.706 ms |             204.0000 |                - | 21000.0000 | 1000.0000 |     - |  169.3 MB |